nextgens: "What about using 90L here instead of a cast ?"
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@19605 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/NodeStats.java
@@ -585,7 +585,7 @@ public class NodeStats implements Persistable {
 			successfulSskOfferReplyBytesReceivedAverage.currentValue() * numSSKOfferReplies;
 		}
 		double bandwidthAvailableInput =
-			(long)node.getInputBandwidthLimit() * 90; // 90 seconds at full power; avoid integer overflow
+			node.getInputBandwidthLimit() * 90L; // 90 seconds at full power; avoid integer overflow
 		if(bandwidthAvailableInput < 0){
 			Logger.error(this, "Negative available bandwidth: "+bandwidthAvailableInput+" node.ibwlimit="+node.getInputBandwidthLimit()+" node.obwlimit="+node.getOutputBandwidthLimit()+" node.inputLimitDefault="+node.inputLimitDefault);
 		}
